Finished some TODOs and cleaned up the code.
